add BoardFooter component with add column button
